finished setting up chunk mode rendering

+++ app/src/main/java/com/door43/translationstudio/targettranslations/TargetTranslationDetailActivityListener.java
@@ -8,5 +8,5 @@ public interface TargetTranslationDetailActivityListener {
      *
      * @param scrollProgress the percent progress of the scroll
      */
-    void onScrollProgressUpdate(float scrollProgress);
+    void onScrollProgressUpdate(int scrollProgress);
 }
